Update incorporating changes in NPL for review comments - part6
- Added NPL to unsupportedFeaturesOnWindows
- Moved ParsePortsRange to agent/util/parse_port.go
