+++ cmd/antrea-agent/agent.go
@@ -36,7 +36,7 @@ import (
 	"github.com/vmware-tanzu/antrea/pkg/agent/flowexporter/flowrecords"
 	"github.com/vmware-tanzu/antrea/pkg/agent/interfacestore"
 	"github.com/vmware-tanzu/antrea/pkg/agent/metrics"
-	npl "github.com/vmware-tanzu/antrea/pkg/agent/nplagent"
+	npl "github.com/vmware-tanzu/antrea/pkg/agent/npl"
 	"github.com/vmware-tanzu/antrea/pkg/agent/openflow"
 	"github.com/vmware-tanzu/antrea/pkg/agent/proxy"
 	"github.com/vmware-tanzu/antrea/pkg/agent/querier"
