+++ pkg/agent/util/parse_port.go
@@ -1,6 +1,4 @@
-// +build !windows
-
-// Copyright 2020 Antrea Authors
+// Copyright 2019 Antrea Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,36 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package lib
+package util
 
 import (
 	"fmt"
-	"reflect"
 	"strconv"
 	"strings"
-
-	"golang.org/x/sys/unix"
-	"k8s.io/klog"
 )
 
-// HasElem checks if a slice contains an element
-func HasElem(s interface{}, elem interface{}) bool {
-	arrV := reflect.ValueOf(s)
-
-	if arrV.Kind() == reflect.Slice {
-		for i := 0; i < arrV.Len(); i++ {
-			// panics if slice element points to an unexported struct field
-			// see https://golang.org/pkg/reflect/#Value.Interface
-			if arrV.Index(i).Interface() == elem {
-				return true
-			}
-		}
-	}
-
-	return false
-}
-
-// ParsePortsRange parses port range and check if valid
+// ParsePortsRange parses port range and checks if valid.
 func ParsePortsRange(portRangeConfig string) (start, end int, err error) {
 	portsRange := strings.Split(portRangeConfig, "-")
 	if len(portsRange) != 2 {
@@ -63,25 +40,4 @@ func ParsePortsRange(portRangeConfig string) (start, end int, err error) {
 	}
 
 	return start, end, nil
-}
-
-// IsPortAvailable checks if a port is free or being used by any other process
-func IsPortAvailable(mPort int) bool {
-	fd, err := unix.Socket(unix.AF_INET, unix.SOCK_STREAM, unix.IPPROTO_TCP)
-	if err != nil {
-		klog.Warningf("unix socket creation failed with error: %v", err)
-		return false
-	}
-	defer unix.Close(fd)
-
-	err = unix.Bind(fd, &unix.SockaddrInet4{
-		Port: mPort,
-		Addr: [4]byte{0, 0, 0, 0},
-	})
-
-	if err != nil {
-		return false
-	}
-
-	return true
 }
